+++ website/src/pages/index.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default () => (
+  <div>Did Doug Ford really win the 2018 Ontario CPC election? If so how?</div>
+);